+++ shardeumd/mempool.go
@@ -54,6 +54,7 @@ func (app *ShardeumApp) configureEVMMempool(appOpts servertypes.AppOptions, logg
 		),
 	)
 	app.SetPrepareProposal(abciProposalHandler.PrepareProposalHandler())
+	app.SetProcessProposal(abciProposalHandler.ProcessProposalHandler())
 
 	return nil
 }